Move bits to another module
