+++ .gitignore
@@ -3,3 +3,4 @@
 Cargo.lock
 *.bpc
 *.mm
+/.vscode
